move src/read.rs -> src/io/read.rs
