+++ tokio-io/src/io/mod.rs
@@ -9,11 +9,13 @@
 //! [found online]: https://tokio.rs/docs/getting-started/core/
 //! [low level details]: https://tokio.rs/docs/going-deeper-tokio/core-low-level/
 
+mod read;
+
 pub use allow_std::AllowStdIo;
 pub use copy::{copy, Copy};
 pub use flush::{flush, Flush};
 pub use lines::{lines, Lines};
-pub use read::{read, Read};
+pub use self::read::{read, Read};
 pub use read_exact::{read_exact, ReadExact};
 pub use read_to_end::{read_to_end, ReadToEnd};
 pub use read_until::{read_until, ReadUntil};
